requirements.txt
django<2
django-tastypie
psycopg2-binary
